// src/app/maps/interfaces/places.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
  type:        string;
  query:       string[];
  features:    Feature[];
  attribution: string;
}

export interface Feature {
  id:                   string;
  type:                 string;
  place_type:           string[];
  relevance:            number;
  properties:           Properties;
  text:                 string;
  place_name:           string;
  center:               number[];
  geometry:             Geometry;
  context:              Context[];
  matching_text?:       string;
  matching_place_name?: string;
  bbox?:                number[];
}

export interface Context {
  id:          string;
  text:        string;
  wikidata?:   string;
  short_code?: string;
}

export interface Geometry {
  coordinates: number[];
  type:        string;
}

export interface Properties {
  foursquare?: string;
  landmark?:   boolean;
  address?:    string;
  category?:   string;
  accuracy?:   string;
  wikidata?:   string;
}